requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
psycopg2
requests
pytest
django_extensions
drf_yasg